+++ lib/weixin_js_sdk/ticket.rb
@@ -2,21 +2,25 @@ module WeixinJsSDK
   class Ticket
     URI_TEMPLATE = 'https://api.weixin.qq.com/cgi-bin/ticket/getticket?access_token=%{access_token}&type=jsapi'.freeze
 
-    def initialize(access_token: '')
-      @access_token = access_token
-    end
+    attr_reader :token, :expires_in
 
-    def fetch
+    def initialize(access_token: '')
       url = URI_TEMPLATE % {
-        access_token: @access_token
+        access_token: access_token
       }
 
-      json = Util.get_json(url)
+      json = Util.get_json(url) rescue {}
 
-      ticket = json['ticket']
-      expires_in = json['expires_in']
-
-      ticket
+      @token = json['ticket']
+      @expires_in = json['expires_in']
     end
+
+    # <b>DEPRECATED:</b> Please use <tt>token</tt> instead.
+    def fetch
+      warn "[DEPRECATION] `fetch` is deprecated.  Please use `token` instead."
+      token
+    end
+
+    alias_method :ticket, :token
   end
 end