+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## master
 
+## v0.0.2
+
+* Deprecated `fetch`
+* Add `token` and `expires_in` to WeixinJsSDK::AccessToken
+* Add `token` and `expires_in` to WeixinJsSDK::Ticket
+
 ## v0.0.1
 
 * Initial release